added client tester for node connections

+++ FileSystem/src/structs/node.h
@@ -7,7 +7,7 @@
 typedef struct {
 	char *id;
 	t_bitarray *blocks;
-	size_t blocksCount;
+	uint32_t blocksCount;
 } node_t;
 
 bson_t* node_getBSON(node_t *node);
@@ -20,7 +20,7 @@ bool node_isBlockUsed(node_t* node, off_t blockIndex);
 int node_getBlocksFreeCount(node_t *node);
 off_t node_getFirstFreeBlock(node_t *node);
 
-node_t* node_create(size_t blocksCount);
+node_t* node_create(uint32_t blocksCount);
 void node_free(node_t* node);
 
 void node_printBlocksStatus(node_t *node);
